Fix: Extract AbstractConfig
